+++ XamarinApp/LAMA/LAMA/LAMA/Models/InventoryItemStorage.cs
@@ -9,7 +9,7 @@ namespace LAMA.Models
 
         //{ "ID", "name", "description", "taken", "free", "takenBy" };
         [PrimaryKey]
-        public int ID { get; set; }
+        public long ID { get; set; }
         public string name { get; set; }
         public string description { get; set; }
         public string taken { get; set; }
@@ -19,7 +19,7 @@ namespace LAMA.Models
         public long lastChange { get; set; }
         public void makeFromStrings(string[] input)
         {
-            ID = Helpers.readInt(input[0]);
+            ID = Helpers.readLong(input[0]);
             name = input[1];
             description = input[2];
             taken = input[3];
